requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-chroma==0.1.0
langchain-huggingface==0.0.1
langchain-groq==0.0.1
chromadb==0.4.22
sentence-transformers==2.2.2
gitpython==3.1.40
pydantic==2.5.0
python-multipart==0.0.6